[#55] refactor : atomsKey.ts 생성 및 분리
