+++ src/components/BottomSheetModal/ItemBrandSelectModalComponent/BrandList.tsx
@@ -4,11 +4,9 @@ import React from 'react'
 import { useSetRecoilState } from 'recoil'
 import useModals from '../../Modals/hooks/useModals'
 import { modals } from '../../Modals'
-import {
-  Brand,
-  selectedBrandState,
-} from '../../../pages/item/create/components/BrandItemField/BrandItemField'
+import {} from '../../../pages/item/create/components/BrandItemField/BrandItemField'
 import BrandLogo from '../../BrandLogo/BrandLogo'
+import { Brand, selectedBrandState } from '../../../config/atomKeys'
 
 const BrandList = () => {
   const brandList = [
