+++ src/components/BottomSheetModal/ItemBrandSelectModalComponent/HotBrand.tsx
@@ -6,10 +6,8 @@ import useModals from '../../Modals/hooks/useModals'
 import { modals } from '../../Modals'
 import Chip from '../../Chip/Chip'
 import { ChipWrapper } from '../ItemBrandSelectModal'
-import {
-  Brand,
-  selectedBrandState,
-} from '../../../pages/item/create/components/BrandItemField/BrandItemField'
+import {} from '../../../pages/item/create/components/BrandItemField/BrandItemField'
+import { Brand, selectedBrandState } from '../../../config/atomKeys'
 
 const HotBrand = () => {
   const brandList = [
