+++ src/config/atomKeys.ts
@@ -1,5 +1,42 @@
 import { atom } from 'recoil'
 
+// 날짜, 장소 Atoms //
+export const selectedDateState = atom<Date | undefined>({
+  // API 호출 시 null로 변환해서 전달
+  key: 'selectedDateState',
+  default: undefined,
+})
+
+export const selectedPlaceState = atom<string>({
+  key: 'selectedPlaceState',
+  default: '',
+})
+
+// 브랜드, 제품명 Atoms //
+export interface Brand {
+  id?: number
+  brandKr?: string
+  brandEn?: string
+  brandImgUrl?: string
+}
+
+export const selectedBrandState = atom<Brand>({
+  key: 'selectedBrand',
+  default: {},
+})
+
+export const itemNameState = atom<string>({
+  key: 'itemName',
+  default: '',
+})
+
+// 아이템 가격 Atoms
+export const itemPriceState = atom<number | undefined>({
+  key: 'itemPriceState',
+  default: 0,
+})
+
+// 추가정보 페이지 Atoms //
 export const addInfoTextState = atom<string>({
   key: 'addInfoTextState',
   default: '',
